+++ scrape/dagster.yaml
@@ -1,0 +1,37 @@
+storage:
+  postgres:
+    postgres_db:
+      username:
+        env: DAGSTER_POSTGRES_USER
+      password:
+        env: DAGSTER_POSTGRES_PASSWORD
+      hostname:
+        env: DAGSTER_POSTGRES_HOST
+      db_name:
+        env: DAGSTER_POSTGRES_DB
+      port:
+        env: DAGSTER_POSTGRES_PORT
+
+run_launcher:
+  module: dagster.core.launcher
+  class: DefaultRunLauncher
+
+run_coordinator:
+  module: dagster.core.run_coordinator
+  class: DefaultRunCoordinator
+
+compute_logs:
+  module: dagster.core.storage.local_compute_log_manager
+  class: LocalComputeLogManager
+  config:
+    base_dir: "/app/dagster_storage/compute_logs"
+
+
+local_artifact_storage:
+  module: dagster.core.storage.file_manager
+  class: LocalFileManager
+  config:
+    base_dir: "/app/dagster_storage/storage"
+
+code_servers:
+  local_startup_timeout: 60
