+++ scrape/packages/analogdb/Makefile
@@ -1,0 +1,7 @@
+.PHONY: test
+test :
+	uv run pytest .
+
+.PHONY: install
+install :
+	uv pip install -e .
