+++ scrape/packages/analogdb/pyproject.toml
@@ -1,0 +1,19 @@
+[project]
+name = "analogdb"
+version = "0.1.0"
+description = "analogdb API client"
+readme = "README.md"
+requires-python = ">=3.13"
+dependencies = [
+    "requests>=2.32.3",
+    "retry>=0.9.2",
+]
+
+[dependency-groups]
+dev = [
+    "pytest>=8.4.0",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
