+++ scrape/pyproject.toml
@@ -1,28 +1,35 @@
 [project]
-name = "scrape"
+name = "dagster-app"
 version = "0.1.0"
-description = "analogdb dagster scrape pipeline"
-readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "dagster>=1.10.18",
-    "dagster-webserver>=1.10.18",
-    "extcolors>=1.0.0",
-    "numpy>=2.2.6",
-    "openai>=1.84.0",
-    "pillow>=11.2.1",
-    "praw>=7.8.1",
-    "requests>=2.32.3",
-    "retry>=0.9.2",
-    "scipy>=1.15.3",
-    "spacy>=3.8.7",
-    "webcolors>=24.11.1",
+    "dagster>=1.5.0",
+    "dagster-webserver>=1.5.0",
+    "analogdb",
+    "scrape",
+    "python-dotenv>=1.1.0",
 ]
 
-[dependency-groups]
-dev = [
+[tool.uv.workspace]
+members = [
+    "packages/analogdb",
+    "packages/scrape"
+]
+
+[tool.uv]
+dev-dependencies = [
     "pytest>=8.4.0",
+    "black>=23.0.0",
+    "ruff>=0.1.0",
 ]
 
-[tool.hatch.build.targets.wheel]
-packages = ["analogdb"]
+[tool.uv.sources]
+analogdb = { workspace = true }
+scrape = { workspace = true }
+
+[tool.dagster]
+module_name = "dagster_app.definitions"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
